Rename commands
