+++ packages/transaction-pool/source/commands/tx-pool-restart.test.ts
@@ -1,7 +1,7 @@
 import { Identifiers, Services } from "@mainsail/cli";
 
 import { Console, describe } from "../../../test-framework/source";
-import { Command } from "./api-restart";
+import { Command } from "./tx-pool-restart";
 
 describe<{
 	cli: Console;
