+++ packages/transaction-pool/source/commands/tx-pool-stop.test.ts
@@ -1,7 +1,7 @@
 import { Identifiers, Services } from "@mainsail/cli";
 
 import { Console, describe } from "../../../test-framework/source";
-import { Command } from "./api-stop";
+import { Command } from "./tx-pool-stop";
 
 describe<{
 	cli: Console;
